feat(ios): add iPadOS/iOS target and simulator dev environment
Build and run the Tauri shell on iPad with the full Rust slicing engine
on-device.

The app entry point moves into a library crate: iOS has no Rust `main`, so
the generated Xcode project links this crate as a `staticlib` and calls the
symbol `tauri::mobile_entry_point` exports. Desktop keeps a three-line
`main.rs` that calls the same `run()`, so a change cannot silently skip
mobile.

The engine excludes `cli`, `server` and `db` on iOS — a sandboxed app has no
command line and must not bind a listener — which keeps clap, actix-web and
sea-orm/sqlx off the Apple mobile targets (333 packages vs 495 for macOS).
`printer` stays, so sending G-code from an iPad keeps the native, CORS-free
transport.

`Info.ios.plist` declares local-network and ATS entries up front; without
them iOS 14+ blocks all LAN traffic and every Moonraker printer would appear
permanently offline.

Tooling: `ios-doctor.sh` verifies the toolchain and explains each gap
(notably that Command Line Tools ship no iOS SDK, and that `tauri ios dev`
ignores DEVELOPER_DIR so `xcode-select` is the only fix), `ios-simulator.sh`
resolves and boots the largest current iPad, and `ios-dev.sh` ties them
together. `ui-desktop/package.json` gains a `tauri` passthrough script that
the generated Xcode build phase requires.

Verified end-to-end: app builds, installs and runs on an iPad Pro 13-inch
(M5) simulator running iOS 26.5.
