+++ package.json
@@ -12,6 +12,13 @@
     "ui:build:web-slicer": "pnpm --filter slicer-ui build:web-slicer",
     "desktop:dev": "pnpm --filter slicer-ui-desktop dev",
     "desktop:build": "pnpm --filter slicer-ui-desktop build",
+    "ios:doctor": "scripts/ios-doctor.sh",
+    "ios:setup": "scripts/ios-doctor.sh --fix",
+    "ios:init": "scripts/ios-doctor.sh && pnpm --filter slicer-ui-desktop exec tauri ios init",
+    "ios:simulator": "scripts/ios-simulator.sh",
+    "ios:dev": "scripts/ios-dev.sh",
+    "ios:open": "pnpm --filter slicer-ui-desktop exec tauri ios dev --open",
+    "ios:build": "pnpm --filter slicer-ui-desktop exec tauri ios build",
     "build:wasm": "cargo build --lib --target wasm32-unknown-unknown --release && wasm-bindgen target/wasm32-unknown-unknown/release/slicer_engine.wasm --target web --out-dir ui/src/generated/scene-wasm --out-name scene_engine",
     "build:wasm:web-slicer": "cargo build --lib --target wasm32-unknown-unknown --release --features web-slicer && wasm-bindgen target/wasm32-unknown-unknown/release/slicer_engine.wasm --target web --out-dir ui/src/generated/scene-wasm --out-name scene_engine",
     "gen-schemas": "cargo run -- gen-schemas --output-dir ui/src/schemas",
