+++ ui-desktop/src-tauri/capabilities/default.json
@@ -1,21 +1,13 @@
 {
   "$schema": "../gen/schemas/desktop-schema.json",
   "identifier": "default",
-  "description": "Default permissions for the slicer desktop app",
+  "description": "Permissions the slicer UI needs on every platform (desktop and mobile)",
   "windows": ["main"],
   "permissions": [
     "core:default",
     "core:event:allow-listen",
     "core:event:allow-unlisten",
     "core:event:allow-emit",
-    "core:menu:default",
-    "core:window:allow-start-dragging",
-    "core:window:allow-minimize",
-    "core:window:allow-maximize",
-    "core:window:allow-unmaximize",
-    "core:window:allow-toggle-maximize",
-    "core:window:allow-is-maximized",
-    "core:window:allow-close",
     "dialog:allow-open",
     "dialog:allow-save",
     "fs:allow-read-file",
