+++ ui-desktop/src-tauri/gen/apple/Sources/slicer-ui-desktop/bindings/bindings.h
@@ -1,0 +1,8 @@
+#pragma once
+
+namespace ffi {
+    extern "C" {
+        void start_app();
+    }
+}
+
