+++ ui-desktop/src-tauri/gen/apple/Sources/slicer-ui-desktop/main.mm
@@ -1,0 +1,6 @@
+#include "bindings/bindings.h"
+
+int main(int argc, char * argv[]) {
+	ffi::start_app();
+	return 0;
+}
