+++ ui-desktop/src-tauri/tauri.conf.json
@@ -12,13 +12,17 @@
   "bundle": {
     "active": true,
     "targets": ["dmg", "app", "msi", "nsis"],
+    "category": "GraphicsAndDesign",
     "icon": [
       "icons/32x32.png",
       "icons/128x128.png",
       "icons/128x128@2x.png",
       "icons/icon.icns",
       "icons/icon.ico"
-    ]
+    ],
+    "iOS": {
+      "minimumSystemVersion": "16.0"
+    }
   },
   "app": {
     "security": {
